+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Each individual change should have a link to the pull request after the descript
 Changed
 ^^^^^^^
 - DatetimeInfoExtractor.mappings_provided changed from a dict.keys() object to list so transformer is serialisable. `#258 <https://github.com/lvgig/tubular/pull/258>`_
+- Created BaseNumericTransformer class to support test refactor of numeric file
 
 1.3.0 (2024-06-13)
 ------------------